Move functions from type to std_types
